Add files via upload
Direct Marketing Campaigns of a Portuguese Banking Institution

The data that is being worked on is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be (or not) subscribed. 

The project objective was to address a classification problem in order to predict if the client will subscribe to a term deposit (variable y). Output variable Y, which denotes if the client has subscribed to a term deposit based on the direct marketing campaign shows that 88.5% did not materialize into a positive outcome while 11.5% eventually subscribed to a term deposit post the campaign.

Based on initial analysis, it suggests the following:
•	If the client has a credit default, it is not likely that it will yield positive outcome from the marketing campaign
•	Positive outcome was more prominent in instances where the cellular information of client is available
•	The duration of the call has a tendency to be longer in instances where there is a positive outcome from the marketing campaign

It was also observed that the dataset contained various categorical features and imbalance of the output proportion which several techniques were used to preprocess the data.

The best performed model is logistic regression and the worst performed model is neural network which may be a little bit weird because logistic regression is simple but performs better while the neural networks is complicated but performs worse. The reason is that neural networks are more difficult to tune parameters and the best performed parameter set for neural networks have not been found and the second reason is that we have only 4521 observations which is not enough for training deep learning models.

In conclusion, this dataset has two challenges: imbalanced distribution of the dependent variable and mixed type objects. We tried 7 ways including Random Forest, Decision Tree with Adaboost, SVM, KNN, Logistic Regression, Feedforward Neural Network and Naive Bayes Classifier. 

In the future, there are two improvements to increase prediction accuracy:
① To combine all possible parameter sets and models 
②To apply feature selection in neural networks to decrease the input size.




1. Direct Marketing Campaigns of a Portuguese Banking Institution

The data that is being worked on is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be (or not) subscribed. For the purpose of this project, the bank.csv dataset, which consists of 10% (4521 samples) of the full dataset was used. 

The project objective was to address a classification problem in order to predict if the client will subscribe to a term deposit (variable y).


Number of Attributes: 16 + output attribute.

Input variables:
1 - age (numeric)
2 - job : type of job (categorical: "admin.","unknown","unemployed","management","housemaid","entrepreneur","student", "blue-collar","self-employed","retired","technician","services") 
3 - marital : marital status (categorical: "married","divorced","single"; note: "divorced" means divorced or widowed)
4 - education (categorical: "unknown","secondary","primary","tertiary")
5 - default: has credit in default? (binary: "yes","no")
6 - balance: average yearly balance, in euros (numeric) 
7 - housing: has a housing loan? (binary: "yes","no")
8 - loan: has personal loan? (binary: "yes","no") related with the last contact of the current campaign:
9 - contact: contact communication type (categorical: "unknown","telephone","cellular") 
10 - day: last contact day of the month (numeric)
11 - month: last contact month of year (categorical: "jan", "feb", "mar", ..., "nov", "dec")
12 - duration: last contact duration, in seconds (numeric)
13 - campaign: number of contacts performed during this campaign and for this client (numeric, includes last contact)
14 - pdays: number of days that passed by after the client was last contacted from a previous campaign (numeric, -1 means client was not previously contacted)
15 - previous: number of contacts performed before this campaign and for this client (numeric)
16 - poutcome: outcome of the previous marketing campaign (categorical: "unknown","other","failure","success")

Output variable (desired target):
17 - y - has the client subscribed to a term deposit? (binary: "yes","no")

Missing Attribute Values: None